Add rector and PHP8 everything

+++ src/CoreShop/Component/Order/Transformer/OrderDocumentItemTransformerInterface.php
@@ -24,7 +24,7 @@ interface OrderDocumentItemTransformerInterface
         OrderDocumentInterface $orderDocument,
         OrderItemInterface $orderItem,
         OrderDocumentItemInterface $documentItem,
-        float $quantity,
+        int $quantity,
         array $options = []
     ): OrderDocumentItemInterface;
 }
